requirements.txt
azure_storage==0.37.0
beautifulsoup4==4.12.3
numpy==1.26.4
pandas==2.2.2
psycopg2==2.9.9
pyarrow==16.1.0
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.30